% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsos.R
\name{lsos}
\alias{lsos}
\title{List objects and their sizes.}
\usage{
lsos(..., n = 10)
}
\arguments{
\item{...}{Arguments passed to the (internal) helper function,
\code{.ls.objects()}.}

\item{n}{A single integer. If positive, the number of rows of the returned
object. If negative, all but the n last/first number of rows of the object
are returned.}
}
\value{
A \code{base::\link[base]{data.frame}} containing the objects' names,
types, sizes, 'pretty' sizes, number of rows, and number of columns. If the
object is list-like (e.g., an S4 object), then the 'number of rows' is given
by the \code{base::\link[base]{length}} of the object and the 'number of
columns' is set to \code{NA}.
}
\description{
An improved \code{base::\link[base]{ls}} function to list
(and optionally sort) the largest objects in the workspace.
}
\examples{
\dontrun{
lsos()
}

}
\author{
Based on Dirk Eddelbuetal's post to Stack Overflow
(\url{http://stackoverflow.com/q/1358003}) in which he in turn credits Petr
Pikal and David Hinds from the r-help mailing list circa 2004.
}
\seealso{
The function wraps the output of \code{base::\link[base]{ls}} and
\code{pryr::\link[pryr]{object_size}} into a pretty format.
}

