% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairsWithCor.R
\name{pairsWithCor}
\alias{pairsWithCor}
\title{Scatterplot Matrices with Correlation}
\usage{
pairsWithCor(x, use = "everything", method = c("pearson", "kendall",
  "spearman"), ...)
}
\arguments{
\item{x}{the coordinates of points given as numeric columns of a
    matrix or data frame.  Logical and factor columns are converted to
    numeric in the same way that \code{\link{data.matrix}} does.
  }

\item{use}{an optional character string giving a
    method for computing covariances in the presence
    of missing values.  This must be (an abbreviation of) one of the strings
    \code{"everything"}, \code{"all.obs"}, \code{"complete.obs"},
    \code{"na.or.complete"}, or \code{"pairwise.complete.obs"}.}

\item{method}{a character string indicating which correlation
    coefficient (or covariance) is to be computed.  One of
    \code{"pearson"} (default), \code{"kendall"}, or \code{"spearman"}:
    can be abbreviated.}

\item{...}{
    arguments to be passed to or from methods.

    Also, \link{graphical parameters} can be given as can arguments to
    \code{plot} such as \code{main}.  \code{par("oma")} will be set
    appropriately unless specified.
  }
}
\description{
A wrapper around `graphics::pairs()` that includes correlation estimates on
the lower diagonal and a [stats::lowess()] smooth of the scatterplots on the
upper triangle with [graphics::panel.smooth()].
}
\details{
The \eqn{ij}th scatterplot contains \code{x[,i]} plotted against
  \code{x[,j]}.  The scatterplot can be customised by setting panel
  functions to appear as something completely different. The
  off-diagonal panel functions are passed the appropriate columns of
  \code{x} as \code{x} and \code{y}: the diagonal panel function (if
  any) is passed a single column, and the \code{text.panel} function is
  passed a single \code{(x, y)} location and the column name.
  Setting some of these panel functions to \code{\link{NULL}} is
  equivalent to \emph{not} drawing anything there.

  The \link{graphical parameters} \code{pch} and \code{col} can be used
  to specify a vector of plotting symbols and colors to be used in the
  plots.

  The \link{graphical parameter} \code{oma} will be set by
  \code{pairs.default} unless supplied as an argument.

  A panel function should not attempt to start a new plot, but just plot
  within a given coordinate system: thus \code{plot} and \code{boxplot}
  are not panel functions.

  By default, missing values are passed to the panel functions and will
  often be ignored within a panel.  However, for the formula method and
  \code{na.action = na.omit}, all cases which contain a missing values for
  any of the variables are omitted completely (including when the scales
  are selected).
  
  Arguments \code{horInd} and \code{verInd} were introduced in \R
  3.2.0. If given the same value they can be used to select or re-order
  variables: with different ranges of consecutive values they can be
  used to plot rectangular windows of a full pairs plot; in the latter
  case \sQuote{diagonal} refers to the diagonal of the full plot.
}
\examples{
pairsWithCor(USJudgeRatings)
}
\references{
Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988)
  \emph{The New S Language}.
  Wadsworth & Brooks/Cole.
}
