% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spwt.R
\name{spwt}
\alias{spwt}
\title{Scatter plot with transparency}
\usage{
spwt(x, y, col = scales::alpha("black", 0.01), pch = 16, cex = 0.5, ...)
}
\arguments{
\item{x, y}{The \code{x} and \code{y} values to be plotted}

\item{col}{The colour transparency to be used}

\item{pch, cex}{See \code{graphics::\link[graphics]{points}}}

\item{...}{Other arguments passed to \code{graphics::\link[graphics]{plot}}}
}
\value{
The plot
}
\description{
An alternative to a \code{graphics::\link[graphics]{smoothScatter}} that
actually plots all points but uses transparency to avoid overplotting.
Recommended to me by Kasper Hansen.
}
\examples{
spwt(rpois(10000, 5) + rnorm(10000, sd = 0.1), rpois(10000, 5) + rnorm(10000, sd = 0.1))

}
\author{
Peter Hickey based on recommendation of Kasper Hansen
}
