% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw.R
\name{sw}
\alias{sw}
\title{Shortcut to sum the widths of a \linkS4class{GenomicRanges} object}
\usage{
sw(x)
}
\arguments{
\item{x}{A \linkS4class{GenomicRanges} object.}
}
\value{
`sum(as.numeric(width(x))`
}
\description{
Shortcut to sum the widths of a \linkS4class{GenomicRanges} object
}
